requirements/base.txt
Django==1.11.6
django-environ==0.4.1
djangorestframework==3.7.1
